Refactor: SQL ingest path and common types to pgmodel/utils
