+++ pkg/pgmodel/ingestor/ingestor.go
@@ -2,45 +2,32 @@
 // Please see the included NOTICE for copyright information and
 // LICENSE for a copy of the license.
 
-package pgmodel
+package ingestor
 
 import (
 	"fmt"
 
+	"github.com/timescale/promscale/pkg/pgmodel/utils"
 	"github.com/timescale/promscale/pkg/prompb"
-)
-
-const (
-	MetricNameLabelName = "__name__"
 )
 
 var (
 	ErrNoMetricName = fmt.Errorf("metric name missing")
 )
 
-// SeriesID represents a globally unique id for the series. This should be equivalent
-// to the PostgreSQL type in the series table (currently BIGINT).
-type SeriesID int64
-
 // inserter is responsible for inserting label, series and data into the storage.
 type inserter interface {
-	InsertNewData(rows map[string][]samplesInfo) (uint64, error)
+	InsertNewData(rows map[string][]utils.SamplesInfo) (uint64, error)
 	CompleteMetricCreation() error
 	Close()
 }
 
 // SeriesCache provides a caching mechanism for labels and series.
 type SeriesCache interface {
-	GetSeries(lset Labels) (SeriesID, error)
-	SetSeries(lset Labels, id SeriesID) error
+	GetSeries(lset utils.Labels) (utils.SeriesID, error)
+	SetSeries(lset utils.Labels, id utils.SeriesID) error
 	NumElements() int
 	Capacity() int
-}
-
-type samplesInfo struct {
-	labels   *Labels
-	seriesID SeriesID
-	samples  []prompb.Sample
 }
 
 // DBIngestor ingest the TimeSeries data into Timescale database.
@@ -76,8 +63,8 @@ func (i *DBIngestor) CompleteMetricCreation() error {
 // returns: map[metric name][]SamplesInfo, total rows to insert
 // NOTE: req will be added to our WriteRequest pool in this function, it must
 //       not be used afterwards.
-func (i *DBIngestor) parseData(tts []prompb.TimeSeries, req *prompb.WriteRequest) (map[string][]samplesInfo, int, error) {
-	dataSamples := make(map[string][]samplesInfo)
+func (i *DBIngestor) parseData(tts []prompb.TimeSeries, req *prompb.WriteRequest) (map[string][]utils.SamplesInfo, int, error) {
+	dataSamples := make(map[string][]utils.SamplesInfo)
 	rows := 0
 
 	for i := range tts {
@@ -88,14 +75,14 @@ func (i *DBIngestor) parseData(tts []prompb.TimeSeries, req *prompb.WriteRequest
 
 		// Normalize and canonicalize t.Labels.
 		// After this point t.Labels should never be used again.
-		seriesLabels, metricName, err := labelProtosToLabels(t.Labels)
+		seriesLabels, metricName, err := utils.LabelProtosToLabels(t.Labels)
 		if err != nil {
 			return nil, rows, err
 		}
 		if metricName == "" {
 			return nil, rows, ErrNoMetricName
 		}
-		sample := samplesInfo{
+		sample := utils.SamplesInfo{
 			seriesLabels,
 			-1, // sentinel marking the seriesId as unset
 			t.Samples,
