+++ pkg/pgmodel/utils/cache.go
@@ -2,7 +2,7 @@
 // Please see the included NOTICE for copyright information and
 // LICENSE for a copy of the license.
 
-package pgmodel
+package utils
 
 import (
 	"fmt"
