+++ pkg/pgmodel/utils/cache_test.go
@@ -1,7 +1,8 @@
 // This file and its contents are licensed under the Apache License 2.0.
 // Please see the included NOTICE for copyright information and
 // LICENSE for a copy of the license.
-package pgmodel
+
+package utils
 
 import (
 	"math"
