+++ pkg/pgmodel/utils/labels.go
@@ -2,7 +2,7 @@
 // Please see the included NOTICE for copyright information and
 // LICENSE for a copy of the license.
 
-package pgmodel
+package utils
 
 import (
 	"encoding/binary"
@@ -16,22 +16,25 @@ import (
 	"github.com/timescale/promscale/pkg/prompb"
 )
 
+// MetricNameLabelKey is the key that denotes a metric's name in any Prometheus metric label_set.
+const MetricNameLabelKey = "__name__"
+
 // Labels stores a labels.Labels in its canonical string representation
 type Labels struct {
-	names      []string
-	values     []string
-	metricName string
+	Names      []string
+	Values     []string
+	MetricName string
 	str        string
 }
 
-var LabelsInterner = sync.Map{}
+var interner = sync.Map{}
 
-// Get the canonical version of a Labels if one exists.
-// input: the string representation of a Labels as defined by getStr()
+// GetLabels fetches the canonical version of labels if one exists.
+// input: the string representation of labels as defined by getStr()
 // This function should not be called directly, use labelProtosToLabels() or
 // LabelsFromSlice() instead.
 func GetLabels(str string) (l *Labels) {
-	val, ok := LabelsInterner.Load(str)
+	val, ok := interner.Load(str)
 	if !ok {
 		return
 	}
@@ -39,13 +42,13 @@ func GetLabels(str string) (l *Labels) {
 	return
 }
 
-// Try to set a Labels as the canonical Labels for a given string
+// SetLabels sets labels as the canonical labels for a given string
 // representation, returning the canonical version (which can be different in
 // the even of multiple goroutines setting labels concurrently).
 // This function should not be called directly, use labelProtosToLabels() or
 // LabelsFromSlice() instead.
 func SetLabels(str string, lset *Labels) *Labels {
-	val, _ := LabelsInterner.LoadOrStore(str, lset)
+	val, _ := interner.LoadOrStore(str, lset)
 	return val.(*Labels)
 }
 
@@ -56,7 +59,7 @@ func LabelsFromSlice(ls labels.Labels) (*Labels, error) {
 		ll[i].Name = ls[i].Name
 		ll[i].Value = ls[i].Value
 	}
-	l, _, err := labelProtosToLabels(ll)
+	l, _, err := LabelProtosToLabels(ll)
 	return l, err
 }
 
@@ -122,8 +125,8 @@ func getStr(labels []prompb.Label) (string, error) {
 	return builder.String(), nil
 }
 
-// labelProtosToLabels converts a prompb.Label to a canonical Labels object
-func labelProtosToLabels(labelPairs []prompb.Label) (*Labels, string, error) {
+// LabelProtosToLabels converts a prompb.Label to a canonical Labels object
+func LabelProtosToLabels(labelPairs []prompb.Label) (*Labels, string, error) {
 	str, err := getStr(labelPairs)
 	if err != nil {
 		return nil, "", err
@@ -132,19 +135,19 @@ func labelProtosToLabels(labelPairs []prompb.Label) (*Labels, string, error) {
 	if labels == nil {
 		labels = new(Labels)
 		labels.str = str
-		labels.names = make([]string, len(labelPairs))
-		labels.values = make([]string, len(labelPairs))
+		labels.Names = make([]string, len(labelPairs))
+		labels.Values = make([]string, len(labelPairs))
 		for i, l := range labelPairs {
-			labels.names[i] = l.Name
-			labels.values[i] = l.Value
-			if l.Name == MetricNameLabelName {
-				labels.metricName = l.Value
+			labels.Names[i] = l.Name
+			labels.Values[i] = l.Value
+			if l.Name == MetricNameLabelKey {
+				labels.MetricName = l.Value
 			}
 		}
 		labels = SetLabels(str, labels)
 	}
 
-	return labels, labels.metricName, err
+	return labels, labels.MetricName, err
 }
 
 // Get a string representation for hashing and comparison
@@ -168,14 +171,14 @@ func (l *Labels) Equal(b *Labels) bool {
 var _ sort.Interface = (*Labels)(nil)
 
 func (l *Labels) Len() int {
-	return len(l.names)
+	return len(l.Names)
 }
 
 func (l *Labels) Less(i, j int) bool {
-	return l.names[i] < l.names[j]
+	return l.Names[i] < l.Names[j]
 }
 
 func (l *Labels) Swap(i, j int) {
-	l.names[j], l.names[i] = l.names[i], l.names[j]
-	l.values[j], l.values[i] = l.values[i], l.values[j]
+	l.Names[j], l.Names[i] = l.Names[i], l.Names[j]
+	l.Values[j], l.Values[i] = l.Values[i], l.Values[j]
 }
