+++ pkg/pgmodel/utils/metrics.go
@@ -1,4 +1,4 @@
-package pgmodel
+package utils
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
@@ -6,28 +6,28 @@ import (
 )
 
 var (
-	duplicateSamples = prometheus.NewCounter(
+	DuplicateSamples = prometheus.NewCounter(
 		prometheus.CounterOpts{
 			Namespace: util.PromNamespace,
 			Name:      "duplicate_samples_total",
 			Help:      "Total number of processed samples which where duplicates",
 		},
 	)
-	duplicateWrites = prometheus.NewCounter(
+	DuplicateWrites = prometheus.NewCounter(
 		prometheus.CounterOpts{
 			Namespace: util.PromNamespace,
 			Name:      "duplicate_writes_total",
 			Help:      "Total number of writes that contained duplicates",
 		},
 	)
-	decompressCalls = prometheus.NewCounter(
+	DecompressCalls = prometheus.NewCounter(
 		prometheus.CounterOpts{
 			Namespace: util.PromNamespace,
 			Name:      "decompress_calls_total",
 			Help:      "Total number of calls to decompress_chunks_after",
 		},
 	)
-	decompressEarliest = prometheus.NewGaugeVec(
+	DecompressEarliest = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Namespace: util.PromNamespace,
 			Name:      "decompress_min_unix_time",
@@ -37,9 +37,9 @@ var (
 
 func init() {
 	prometheus.MustRegister(
-		duplicateSamples,
-		duplicateWrites,
-		decompressCalls,
-		decompressEarliest,
+		DuplicateSamples,
+		DuplicateWrites,
+		DecompressCalls,
+		DecompressEarliest,
 	)
 }
